refactor: moved add-profile-photo-with-crop.tsx to widget: the rule of imports between layers is violated - cannot import data from entities to shared
