+++ src/widgets/add-profile-photo/add-avatar-button/image-cropper/set-canvas-preview.ts
