+++ src/widgets/add-profile-photo/add-profile-photo-with-crop.tsx
@@ -2,9 +2,9 @@ import { UserAvatar } from '@/entities/profile'
 import { useProfile } from '@/entities/profile/model/store/profile-store'
 import ImageOutlineIcon from '@/shared/assets/icons/ImageOutlineIcon'
 import { createImageUrlFromPhotoFile } from '@/shared/lib/utils/create-image-url-from-photo-file'
-import { AddAvatarButton } from '@/shared/ui/add-profile-photo/add-avatar-button/addAvatarButton'
-import { DeleteAvatarButton } from '@/shared/ui/add-profile-photo/delete-avatar-button/delete-avatar-button'
-import { useMyProfile } from '@/shared/ui/add-profile-photo/useMyProfile'
+import { AddAvatarButton } from '@/widgets/add-profile-photo/add-avatar-button/addAvatarButton'
+import { DeleteAvatarButton } from '@/widgets/add-profile-photo/delete-avatar-button/delete-avatar-button'
+import { useMyProfile } from '@/widgets/add-profile-photo/useMyProfile'
 
 import 'react-image-crop/dist/ReactCrop.css'
 
@@ -19,7 +19,7 @@ export const AddProfilePhotoWithCrop = () => {
   const file = localAvatar?.get('file') as File
   const imageUrl = createImageUrlFromPhotoFile(file)
 
-  const avatarUrl = myProfile.profile?.url ?? (imageUrl || '')
+  const avatarUrl = myProfile.avatarUrl ?? (imageUrl || '')
 
   return (
     <div className={'flex flex-col gap-y-6 py-[1.5rem]'}>
@@ -27,9 +27,9 @@ export const AddProfilePhotoWithCrop = () => {
         <UserAvatar bgColor={'bg-Dark-500'} className={`h-full w-full`} src={avatarUrl}>
           <ImageOutlineIcon />
         </UserAvatar>
-        {myProfile.profile?.url && <DeleteAvatarButton />}
+        {myProfile.avatarUrl && <DeleteAvatarButton />}
       </div>
-      <AddAvatarButton profile={myProfile.profile} />
+      <AddAvatarButton profile={myProfile} />
     </div>
   )
 }
