+++ src/widgets/add-profile-photo/delete-avatar-button/hooks/useDeleteAvatar.ts
