+++ src/widgets/add-profile-photo/useMyProfile.ts
@@ -1,9 +1,9 @@
-import { userService } from '@/entities/user/api/user-api'
+import profileApi from '@/entities/profile/api/profile-api'
 import { useQuery } from '@tanstack/react-query'
 
 export const useMyProfile = () => {
   return useQuery({
-    queryFn: async () => await userService.getProfile(),
+    queryFn: async () => await profileApi.getProfile(),
     queryKey: ['myProfile'],
   })
 }
